feat: added builder for external and contrib plugins, removed old component builder
