+++ juturna/components/__init__.py
@@ -5,4 +5,9 @@ from juturna.components._pipeline import Pipeline
 from juturna.components._buffer import Buffer
 
 
-__all__ = ['Message', 'Node', 'Pipeline', 'Buffer']
+__all__ = [
+    'Message',
+    'Node',
+    'Pipeline',
+    'Buffer',
+]
